+++ libs/cardano-ledger-core/src/Cardano/Ledger/Hashes.hs
@@ -39,7 +39,7 @@ module Cardano.Ledger.Hashes (
   EraIndependentTxAuxData,
   EraIndependentPParamView,
   EraIndependentScriptIntegrity,
-  EraIndependentEb,
+  EraIndependentEbReferences,
 
   -- * Hashes
 
@@ -155,7 +155,7 @@ data EraIndependentPParamView
 
 data EraIndependentScriptIntegrity
 
-data EraIndependentEb
+data EraIndependentEbReferences
 
 --------------------------------------------------------------------------------
 -- Signature Hashes