+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/TreeDiff.hs
@@ -351,4 +351,4 @@ instance (ToExpr h, ToExpr (BlockBody era)) => ToExpr (Block h era)
 
 instance ToExpr TestBlockHeader
 
-instance ToExpr EbAnnouncement
+instance ToExpr EbReferencesAnnouncement