+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 1.22.0.0
 
-* Add `EbAnnouncement` (moved from `cardano-protocol`)
-* Add `blockHeaderL`, `blockBodyL`
+* Rename `EraIndependentEb` to `EraIndependentEbReferences`
+* Add `EbReferencesAnnouncement` (moved and renamed from `EbAnnouncement` in `cardano-protocol`)
+* Add `blockHeaderL`, `blockBodyL`, `bhviProtVerL`
 * Add `BlockHeaderVersionInfo` to `Cardano.Ledger.Block`
 * Add `versionInfoBlockHeaderL` to `LeiosEraBlockHeader`
 * Add `ProtocolEraBlockHeader` type family constraint to `EraBlockBody`